Create bespoke existing entity card component
